Improve perf and carry logic.

Compiler    | Before  | After
------------|---------|--------
MSVC x86 %f | 357.983 | 356.370
MSVC x86 %e | 107.619 | 105.532
LLVM x86 %f | 352.767 | 353.259
LLVM x86 %e | 104.455 | 104.600

Compiler    | Before  | After   | Speedup
------------|---------|---------|--------
MSVC x64 %f | 159.110 | 118.643 | 34.1%
MSVC x64 %e |  54.288 |  44.169 | 22.9%
LLVM x64 %f | 116.734 |  97.031 | 20.3%
LLVM x64 %e |  40.285 |  36.193 | 11.3%
LLVM 128 %f | 101.752 |  98.859 | 2.9%
LLVM 128 %e |  37.498 |  37.120 | 1.0%

Note that this slightly improves the "headline perf" of the algorithm. Comparing the previous fastest (LLVM uint128) to the new fastest (LLVM x64 intrinsics), %f is 4.9% faster, and %e is 3.6% faster.

Individual commits:

---

Use bitwise OR in umul128() to improve codegen.

When concatenating two 32-bit values, MSVC's x86 codegen is affected by plus versus bitwise OR.

For plus, the codegen is:

```
add eax, DWORD PTR _b00$1$[esp+16]
adc edi, 0
```

That's an add followed by add-with-carry; i.e. the compiler is emitting a general 64 + 32 addition.

For bitwise OR, the codegen is:

```
or  eax, DWORD PTR _b00$1$[esp+16]
```

Bitwise OR is both conceptually simpler and more efficient.

---

Use bitwise OR in umul256() for consistency.

While Clang x64 uint128 codegen is unaffected, we should still change plus to bitwise OR for consistency and conceptual simplicity.

---

Improve x64 intrinsic perf with the magic multiply.

This ports the magic multiply technique in 7209c36 from uint128 to x64 intrinsics. Interestingly, this is slightly faster for Clang/LLVM (at least on Windows), presumably due to weak uint128 optimizations.

Note that when uint128 and x64 intrinsics are unavailable (e.g. x86), the magic multiply is slower than the technique of calling `mod1e9()` repeatedly, which is why that codepath is being retained here.

`umul256_hi128_lo64()` uses `uint64_t` for all math. Therefore, it can't take advantage of the fact that double-width products can absorb two single digits. So, it has explicit carry checks, and names its variables `temp` instead of `mid`.

---

Improve uint128 carrying perf.

`mulShift()` computes three double-width products (b0, b1, b2). It can be restructured to improve codegen by avoiding explicit carrying.

This takes advantage of the fact (used elsewhere) that double-width products can absorb up to two single digits via addition without overflowing. Here, we have b0, b1, and b2 partially overlapping. The low 64 bits of b0 don't get added to anything (and are ignored, because we shift them away later). The high 64 bits of b0 can be added to b1, without overflow due to the double-width product property. This is stored as mid. (The low 64 bits of mid are similarly ignored.) Finally, the high 64 bits of mid can be added to b2, producing the upper 128 bits of the sum.

---

Improve non-uint128 mulShift() carry logic.

In 57b918f, I saw two additions with only one overflow check for carrying, and added an assertion. Now that I understand what the code is doing, the assertion can be removed - but the overflow check actually needs to be changed.

This code is forming and adding double-width products, represented as separate parts (high0 and low0, etc.). This means that we can take advantage of the double-width product property, but we need to be careful to keep track of what each variable represents. We're adding:

```
               high0  low0
+        high1 low1
+ high2  low2
===========================
  s1high s1low s0high s0low
```

Adding low1 to high0 generates s0high. This is taking a double-width product (high1 low1) and adding a single digit (high0), which can't overflow beyond high1. However, it can internally carry within the double-width value (which happens behind the scenes for uint128 + uint64). Here, we need to manually carry with c1, which will be added to high1 (and that cannot overflow).

Then we take high1 + c1 and add low2, generating s1low. Again this is taking a double-width product (high2 low2) and adding a single digit (high1 + c1), which can't overflow beyond high2. Again it can internally carry, so we need a single check.

Here is where the code must be fixed: we're adding two digits, low2 + (high1 + c1), so we need to compare against one of the inputs. Either low2 or (high1 + c1) would be fine, but comparing against high1 is incorrect. (I believe that the code was getting away with it due to the specific ranges of the inputs.) Since low2 is already available, I'm using that.

Note that I have also changed the computation of c1 for stylistic consistency, although there is no correctness issue there. (In a long chain of overlapping additions, we would want to consistently compare against the low part.)

Thanks again to @abolz for explaining the double-width product property to me: https://github.com/ulfjack/ryu/issues/94#issuecomment-458942435